pyproject.toml
[project]
name = "company-data-generator"
version = "0.1.0"
description = "架空の日本国内の株式会社の社内データを生成するジェネレータ"
readme = "README.md"
authors = [
    { name = "Yusuke Suzuki", email = "yusuke.suzuki@microsoft.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-ai-inference>=1.0.0b9",
    "azure-identity>=1.25.2",
    "click>=8.3.1",
    "gradio>=6.6.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]

[project.scripts]
company-data-generator = "company_data_generator.cli:main"

[build-system]
requires = ["uv_build>=0.10.5,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
